Move type checking notes
